+++ src/index.ts
@@ -20,6 +20,7 @@ import {
 } from './accounts'
 import { walletClientToAccount } from './accounts/walletClient'
 import {
+  getIntentStatus as getIntentStatusInternal,
   getMaxSpendableAmount as getMaxSpendableAmountInternal,
   getPortfolio as getPortfolioInternal,
   sendTransaction as sendTransactionInternal,
@@ -159,7 +160,7 @@ interface RhinestoneAccount {
   getPortfolio: (onTestnets?: boolean) => Promise<Portfolio>
   getMaxSpendableAmount: (
     chain: Chain,
-    tokenAddress: Address,
+    tokenAddress: Address | TokenSymbol,
     gasUnits: bigint,
     sponsored?: boolean,
   ) => Promise<bigint>
@@ -409,20 +410,20 @@ async function createRhinestoneAccount(
   /**
    * Get the maximum spendable token amount on the target chain
    * @param chain Target chain
-   * @param tokenAddress Token address (on the target chain)
+   * @param token Token address (on the target chain)
    * @param gasUnits Gas cost estimate for the transaction execution
    * @returns Maximum spendable amount in absolute units
    */
   function getMaxSpendableAmount(
     chain: Chain,
-    tokenAddress: Address,
+    token: Address | TokenSymbol,
     gasUnits: bigint,
     sponsored: boolean = false,
   ) {
     return getMaxSpendableAmountInternal(
       config,
       chain,
-      tokenAddress,
+      token,
       gasUnits,
       sponsored,
     )
@@ -536,6 +537,10 @@ class RhinestoneSDK {
       paymaster: this.paymaster,
     }
     return createRhinestoneAccount(rhinestoneConfig)
+  }
+
+  getIntentStatus(intentId: bigint) {
+    return getIntentStatusInternal(this.apiKey, this.endpointUrl, intentId)
   }
 }